Se modifica message.php para que no de una alerta sino un mensaje modal. Además, se crea el backend confirmation.php que será el encargado de mostrar un mensaje de confirmación antes de eliminar cualquier registros de la BBDD. Por otro lado se han adaptado los componentes finalizados a esta nueva gestión de mensajes.

+++ php/deleteZone.php
@@ -13,6 +13,13 @@ require "database.php";
 $idZone = $_GET["Id"];
 $zoneName = $_GET["zoneName"];
 
+if (isset($_GET["delete"]) ){
+    $delete = $_GET["delete"];
+} else {
+    $delete = "";
+}
+
+
 try {
     $sql = "SELECT id_reservation  FROM reservations WHERE zone_id = :idzone AND reservation_status = 'A'";
     $query = $conn->prepare($sql); 
@@ -23,25 +30,34 @@ try {
         $_SESSION['successFlag'] = "N";
         $_SESSION['message'] = "No se puede eliminar la zona $zoneName ya que existen reservas activas asociadas a esta zona. Cancela antes las reservas activas y vuelve a intentarlo." ; 
     } else {
-        
-        try {
-            $sql   = "DELETE FROM zones WHERE id_zone = $idZone" ;
-            $count = $conn->exec($sql);  
+        if ($delete == "yes") {
+            try {
+                $sql   = "DELETE FROM zones WHERE id_zone = $idZone" ;
+                $count = $conn->exec($sql);  
 
-            if ($count == 0) {
-                $_SESSION['successFlag'] = "N"; 
-                $_SESSION['message'] = "Ha habido un problema y no se ha podido eliminar la zona $zoneName." ; 
-            } else {
-                $_SESSION['successFlag'] = "Y";
-                $_SESSION['message'] = "La zona $zoneName ha sido eliminada correctamente." ; 
+                if ($count == 0) {
+                    $_SESSION['successFlag'] = "N"; 
+                    $_SESSION['message'] = "Ha habido un problema y no se ha podido eliminar la zona $zoneName." ; 
+                } else {
+                    $_SESSION['successFlag'] = "Y";
+                    $_SESSION['message'] = "La zona $zoneName ha sido eliminada correctamente." ; 
+                }
+            
+            } catch(PDOException $e) {
+                $_SESSION['successFlag'] = "N";
+                $queryError = $e->getMessage();  
+                $_SESSION['message'] = "Se ha detectado un problema a la hora de eliminar la zona $zoneName. </br> Descripción del error: " . $queryError ; 
+            } finally {
+                //Una se haya intentado eliminar una zona, se inicializa la variable de sesión 
+                $_SESSION['idZone'] = "" ;
+                $_SESSION['zoneName'] = "" ;
             }
-                
-           
-
-        } catch(PDOException $e) {
-            $_SESSION['successFlag'] = "N";
-            $queryError = $e->getMessage();  
-            $_SESSION['message'] = "Se ha detectado un problema a la hora de eliminar la zona $zoneName. </br> Descripción del error: " . $queryError ; 
+            
+        } else {
+            $_SESSION['confirmation'] = "";
+            $_SESSION['idZone']   = $idZone;
+            $_SESSION['zoneName'] = $zoneName;
+            $_SESSION['message']  = "Estás a punto de eliminar la zona $zoneName. Esto también eliminará las reservas pasadas asociadas a esa zona." ;
         }
     }
